requirements.txt
aws-cdk-lib==2.206.0
constructs>=10.0.0,<11.0.0
amazon-textract-idp-cdk-constructs
PyJWT
boto3
python-dotenv>=1.0.0
